get axelar deposit address
