+++ src/hooks/useAsset.ts
@@ -10,6 +10,7 @@ import { AssetType, WhiteListType, BalanceListType } from 'types/asset'
 import {
   BlockChainType,
   isIbcNetwork,
+  isAxelarNetwork,
   allowedCoins,
   IbcNetwork,
 } from 'types/network'
@@ -81,9 +82,8 @@ const useAsset = (): {
         let balanceWhiteList = _.map(whiteList, (token) => ({ token }))
         switch (toBlockChain) {
           case BlockChainType.terra:
-            balanceWhiteList = balanceWhiteList.filter(
-              ({ token }): boolean =>
-                token.startsWith('terra1')
+            balanceWhiteList = balanceWhiteList.filter(({ token }): boolean =>
+              token.startsWith('terra1')
             )
             break
           case BlockChainType.ethereum:
@@ -106,27 +106,21 @@ const useAsset = (): {
             break
           default:
             // ibc chain
-            if(isIbcNetwork(toBlockChain)) {
-              balanceWhiteList = balanceWhiteList.filter(
-                ({ token }): boolean =>
-                  token.startsWith('terra1') &&
-                  allowedCoins[toBlockChain as IbcNetwork].includes(token)
-              )
-            // axelar
-            } else {
-              balanceWhiteList = balanceWhiteList.filter(
-                ({ token }): boolean =>
-                  token.startsWith('terra1') &&
-                  allowedCoins[BlockChainType.axelar].includes(token)
-              )
-              console.log(allowedCoins[BlockChainType.axelar])
-            }
+            balanceWhiteList = balanceWhiteList.filter(
+              ({ token }): boolean =>
+                token.startsWith('terra1') &&
+                allowedCoins[
+                  isIbcNetwork(toBlockChain)
+                    ? (toBlockChain as IbcNetwork)
+                    : BlockChainType.axelar
+                ].includes(token)
+            )
         }
         balanceList = await getTerraBalances({
           terraWhiteList: balanceWhiteList,
         })
       } else if (NETWORK.isEtherBaseBlockChain(fromBlockChain)) {
-        switch(fromBlockChain){
+        switch (fromBlockChain) {
           case BlockChainType.ethereum:
             whiteList = ethWhiteList
             break
@@ -139,7 +133,7 @@ const useAsset = (): {
         }
         balanceList = await getEtherBalances({ whiteList })
       } else if (isIbcNetwork(fromBlockChain)) {
-        switch(fromBlockChain){
+        switch (fromBlockChain) {
           case BlockChainType.osmo:
             whiteList = osmoWhiteList
             break
@@ -188,6 +182,15 @@ const useAsset = (): {
         allowed.includes(item.terraToken)
       )
       setAssetList(filteredList)
+    } else if (
+      fromBlockChain === BlockChainType.terra &&
+      isAxelarNetwork(toBlockChain)
+    ) {
+      const allowed = allowedCoins[BlockChainType.axelar] as string[]
+      const filteredList = fromList.filter((item) =>
+        allowed.includes(item.terraToken)
+      )
+      setAssetList(filteredList)
     } else {
       setAssetList(fromList)
     }
@@ -198,7 +201,8 @@ const useAsset = (): {
       const bnBalance =
         typeof balance === 'string' ? new BigNumber(balance) : balance
 
-      return fromBlockChain === BlockChainType.terra || isIbcNetwork(fromBlockChain)
+      return fromBlockChain === BlockChainType.terra ||
+        isIbcNetwork(fromBlockChain)
         ? bnBalance.div(ASSET.TERRA_DECIMAL).dp(6).toString(10)
         : bnBalance
             .div(ASSET.ETHER_BASE_DECIMAL / ASSET.TERRA_DECIMAL)
